+++ src/main/java/weterynarz/View/DoctorView/DoctorWelcomeBox.java
@@ -1,58 +1,67 @@
-package weterynarz.View;
+package weterynarz.View.DoctorView;
 
-import java.awt.BorderLayout;
-import java.awt.CardLayout;
-import java.awt.Color;
-import java.awt.Container;
-import java.awt.event.ItemEvent;
+import java.awt.*;
+import java.awt.event.*;
+import java.text.SimpleDateFormat;
 
-import javax.swing.JButton;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.UIManager;
-import javax.swing.UnsupportedLookAndFeelException;
+import javax.swing.*;
 
-import weterynarz.View.PatientView.PatientWelcomeBox;
+import java.util.Date;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
 
-public class TestWelcomeBox 
+public class DoctorWelcomeBox
 {
 	JPanel card;
-    JLabel label, label2, label1;
-    JButton showAnimalsButton;
+    JLabel label, labelx, label2, label1;
+    JButton daneosobowe, alistofanim, lecznice, calendar;
     JFrame frame;
     
-    public TestWelcomeBox()
+    public DoctorWelcomeBox()
     {
-    	frame = new JFrame("Witaj!");
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); 
+    	//Create and set up the window.
+        frame = new JFrame("Witaj!");
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+         
         //Create and set up the content pane.
         addComponentToPane(frame.getContentPane());
        
         //Display the window.
         //frame.pack();
-        
         frame.setSize(290,330);
         frame.setResizable(false);
-        frame.setVisible(false);
+      //  frame.setVisible(false);
     }
      
     public void addComponentToPane(Container pane) 
     {
       
-              
+    	DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
+    	Date date = new Date();
         card = new JPanel();
         card.setBackground(new Color(220, 236, 237));
         //card.setLayout(new BorderLayout());
-        showAnimalsButton = new JButton("Moje zwierz¹tka");
-        JButton add = new JButton("Dodaj zwierzaka");
+        calendar = new JButton("           Moj Terminarz            ");
+        // calendar.setBounds(20, 100, 40, 40);
+        // calendar.setBorder(null);
+
+        daneosobowe = new JButton("     Moje Dane Osobowe     ");
+        alistofanim = new JButton("Zwierzeta pod moja opieka");
+        lecznice = new JButton("   Lista dostepnych lecznic  ");
                   
-        label = new JLabel("Witamy"
-        		+ " na stronie g³ównej.          ");
+        label = new JLabel("Witamy na stronie glownej!");
+        labelx = new JLabel(dateFormat.format(date));
+       // labelx.setLocation(200, 200);
         
         card.add(label);
-        card.add(showAnimalsButton);
-        card.add(add);
+       
+        card.add(daneosobowe);
+        card.add(calendar);
+        card.add(alistofanim);
+        card.add(lecznice);
+        
+        card.add(labelx);
         
         
        // Create the panel that contains the "cards".
@@ -65,21 +74,6 @@ public class TestWelcomeBox
         CardLayout cl = (CardLayout)(card.getLayout());
         cl.show(card, (String)evt.getItem());
     }
-    
-    public void setVisibility(boolean visibility)
-    {
-    	this.frame.setVisible(visibility);
-    }
-    
-    public void setText(String text)
-    {
-    	this.label.setText(text);
-    }
-    
-    public JButton getExampleButton()
-    {
-    	return showAnimalsButton;
-    }
      
 
     private static void createAndShow() {
@@ -87,7 +81,7 @@ public class TestWelcomeBox
         JFrame frame = new JFrame("Witaj!");
         frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); 
         //Create and set up the content pane.
-        PatientWelcomeBox demo = new PatientWelcomeBox();
+        DoctorWelcomeBox demo = new DoctorWelcomeBox();
         demo.addComponentToPane(frame.getContentPane());
        
          
@@ -98,6 +92,31 @@ public class TestWelcomeBox
         frame.setSize(290,330);
         frame.setResizable(false);
         frame.setVisible(true);
+    }
+    
+    public JButton daneOsobowe()
+    {
+    	return daneosobowe;
+    }
+    
+    public JButton aListOfAnim()
+    {
+    	return alistofanim;
+    }
+    
+    public JButton Lecznice()
+    {
+    	return lecznice;
+    }
+    
+    public JButton Calendar()
+    {
+    	return calendar;
+    }
+    
+    public void setVisibility(boolean visibility)
+    {
+    	this.frame.setVisible(visibility);
     }
      
     public static void main(String[] args) {
@@ -125,4 +144,7 @@ public class TestWelcomeBox
             }
         });
     }
+	
+
 }
+
